# server/public/openapi.yaml
openapi: 3.1.0
info:
  title: TeamSpark Pachinko API
  version: 1.0.0
  description: |
    HTTP API for policy configuration, message and alert review, analytics,
    Arcade Engine webhooks, retention, and application settings.
servers:
  - url: http://localhost:3000/api/v1
    description: API base (Next.js app)
...
components:
  schemas:
    Meta:
      type: object
      properties:
        apiVersion:
          type: string
          enum: ["1.0"]
        status:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Status message
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        remaining:
          type: integer
        hasMore:
          type: boolean
        nextCursor:
          type: integer
          nullable: true
        limit:
          type: integer
        sort:
          type: string
          enum: [asc, desc]
    TimeSeriesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
                format: date-time
                description: The timestamp for this data point
              counts:
                type: object
                additionalProperties:
                  type: number
                description: Map of dimension values to their counts at this timestamp
        query:
          type: object
          properties:
            dimension:
              type: string
              description: The dimension being queried
            timeUnit:
              type: string
              enum: [hour, day, week, month]
              description: The time unit used for grouping
            timeRange:
              type: object
              properties:
                start:
                  type: string
                  format: date-time
                  description: Start of the time range
                end:
                  type: string
                  format: date-time
                  description: End of the time range
            filters:
              type: object
              properties:
                serverName:
                  type: string
                serverId:
                  type: integer
                userId:
                  type: string
                clientId:
                  type: integer
                clientType:
                  type: string
                method:
                  type: string
                toolName:
                  type: string
                sourceIP:
                  type: string
...
    Alert:
      type: object
      properties:
        alertId:
          type: integer
          description: Unique identifier for the alert
        messageId:
          type: integer
          description: ID of the message that triggered this alert
        policyId:
          type: integer
          description: ID of the policy that triggered this alert
        filterName:
          type: string
          description: Name of the filter that triggered this alert
        origin:
          type: string
          enum: [client, server]
          description: Origin of the message that triggered this alert
        redactions:
          type: object
          nullable: true
          description: Any redacted fields from the message
        timestamp:
          type: string
          format: date-time
          description: When the alert was triggered
        createdAt:
          type: string
          format: date-time
          description: When the alert was created
        seenAt:
          type: string
          format: date-time
          nullable: true
          description: When the alert was marked as seen
        policyName:
          type: string
          description: Name of the policy that triggered this alert
        policySeverity:
          type: integer
          description: Severity level of the policy
        serverId:
          type: integer
          description: ID of the server that generated the message
        clientId:
          type: integer
          description: ID of the client that generated the message
        clientType:
          type: string
          description: Type of the client that generated the message
      required:
        - alertId
        - messageId
        - policyId
        - filterName
        - timestamp
        - createdAt
        - policyName
        - policySeverity
        - serverId
        - clientId
        - clientType
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              meta:
                $ref: '#/components/schemas/Meta'
...
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            type: object
            properties:
              meta:
                $ref: '#/components/schemas/Meta'
paths:
...
  /alerts:
    get:
      tags: [Alerts]
      summary: List alerts
      parameters:
        - name: sort
          in: query
          schema: { type: string, enum: [asc, desc] }
        - name: cursor
          in: query
          schema: { type: integer }
        - name: messageId
          in: query
          schema: { type: integer }
        - name: policyId
          in: query
          schema: { type: integer }
        - name: conditionName
          in: query
          schema: { type: string }
        - name: seen
          in: query
          schema: { type: boolean }
        - name: severity
          in: query
          schema: { type: integer }
        - name: startTime
          in: query
          schema: { type: string, format: date-time }
        - name: endTime
          in: query
          schema: { type: string, format: date-time }
        - name: source
          in: query
          schema: { type: string }
        - name: payloadToolkit
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: "#/components/schemas/Meta"
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /analytics/alerts/timeSeries:
    get:
      tags: [Analytics]
      summary: Alert time series
      parameters:
        - name: dimension
          in: query
          required: true
          schema:
            type: string
            enum: [policyId, conditionName, seen, severity, source, payloadToolkit]
        - name: timeUnit
          in: query
          required: true
          schema: { type: string, enum: [hour, day, week, month] }
        - name: policyId
          in: query
          schema: { type: integer }
        - name: conditionName
          in: query
          schema: { type: string }
        - name: seen
          in: query
          schema: { type: boolean }
        - name: severity
          in: query
          schema: { type: integer }
        - name: startTime
          in: query
          schema: { type: string, format: date-time }
        - name: endTime
          in: query
          schema: { type: string, format: date-time }
        - name: source
          in: query
          schema: { type: string }
        - name: payloadToolkit
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: "#/components/schemas/Meta"
                  timeSeries:
                    $ref: "#/components/schemas/TimeSeriesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /analytics/alerts/aggregate:
    get:
      tags: [Analytics]
      summary: Alert aggregates
      parameters:
        - name: dimension
          in: query
          required: true
          schema:
            type: string
            enum: [policyId, conditionName, seen, severity, source, payloadToolkit]
        - name: policyId
          in: query
          schema: { type: integer }
        - name: conditionName
          in: query
          schema: { type: string }
        - name: seen
          in: query
          schema: { type: boolean }
        - name: severity
          in: query
          schema: { type: integer }
        - name: startTime
          in: query
          schema: { type: string, format: date-time }
        - name: endTime
          in: query
          schema: { type: string, format: date-time }
        - name: source
          in: query
          schema: { type: string }
        - name: payloadToolkit
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: "#/components/schemas/Meta"
                  aggregate:
                    type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
...